Added CORS Headers and Handling of Options request
Issue: Preflight requests return a 403 since the options method was not being handled. 

I believe this creates the appropriate CORS headers and then handles the options method. Feel free to close if this is not something you would like added, or if there is more involved that i'm not aware of that I should try and add let me know!

Used as Reference: https://enable-cors.org/server_expressjs.html